% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_key_merging.R
\name{complex_key_merging}
\alias{complex_key_merging}
\title{Function that merge based on complex key}
\usage{
complex_key_merging(dp_dat1, dp_dat2, sap_complex_key1, sap_complex_key2)
}
\description{
This function allows you to merge 2 data frames using complex keys
Include options for all.x, all.y and all
}
\keyword{merging}
